+++ man-roxygen/section_progress_bars.R
@@ -1,0 +1,5 @@
+#' @section Progress Bars:
+#' This function supports progress bars via the package \CRANpkg{progressr}.
+#' Simply wrap the function in [progressr::with_progress()] to enable them.
+#' We recommend to use package \CRANpkg{progress} as backend; enable with
+#' `progressr::handlers("progress")`.