+++ tests/testthat/test_parallel.R
@@ -3,6 +3,7 @@ context("parallelization")
 test_that("parallel resample", {
   skip_if_not_installed("future")
   skip_if_not_installed("future.callr")
+  skip_if_not_installed("progressr")
 
   with_future(future.callr::callr, {
     expect_true(use_future())
@@ -10,7 +11,9 @@ test_that("parallel resample", {
     task = tsk("iris")
     learner = lrn("classif.rpart")
 
-    rr = resample(task, learner, rsmp("cv", folds = 3))
+    progressr::with_progress({
+      rr = resample(task, learner, rsmp("cv", folds = 3))
+    })
     expect_resample_result(rr)
     expect_data_table(rr$errors, nrows = 0L)
   })
@@ -26,7 +29,9 @@ test_that("parallel benchmark", {
     task = tsk("iris")
     learner = lrn("classif.rpart")
 
-    bmr = benchmark(benchmark_grid(task, learner, rsmp("cv", folds = 3)))
+    progressr::with_progress({
+      bmr = benchmark(benchmark_grid(task, learner, rsmp("cv", folds = 3)))
+    })
     expect_benchmark_result(bmr)
     expect_equal(bmr$aggregate(conditions = TRUE)$warnings, 0L)
     expect_equal(bmr$aggregate(conditions = TRUE)$errors, 0L)